Rewrite using ffmpeg-python instead of manual command building
Conceptually the idea is very similar. We go through collecting data that will
serve as command-line arguments. The primary difference in ffmpeg-python is the
distinction between input streams, filters, and option flags.

ffmpeg-python creates a model of the input graph and rearranges the commandline
input to be in correct order. This lets us change the order in a few places for
better readability.